+++ vulkano/src/buffer/sys.rs
@@ -298,6 +298,8 @@ impl UnsafeBuffer {
 unsafe impl VulkanObject for UnsafeBuffer {
     type Object = vk::Buffer;
 
+    const TYPE: vk::DebugReportObjectTypeEXT = vk::DEBUG_REPORT_OBJECT_TYPE_BUFFER_EXT;
+
     #[inline]
     fn internal_object(&self) -> vk::Buffer {
         self.buffer