+++ vulkano/src/command_buffer/pool/sys.rs
@@ -225,6 +225,8 @@ unsafe impl DeviceOwned for UnsafeCommandPool {
 unsafe impl VulkanObject for UnsafeCommandPool {
     type Object = vk::CommandPool;
 
+    const TYPE: vk::DebugReportObjectTypeEXT = vk::DEBUG_REPORT_OBJECT_TYPE_COMMAND_POOL_EXT;
+
     #[inline]
     fn internal_object(&self) -> vk::CommandPool {
         self.pool
@@ -246,6 +248,8 @@ pub struct UnsafeCommandPoolAlloc(vk::CommandBuffer);
 
 unsafe impl VulkanObject for UnsafeCommandPoolAlloc {
     type Object = vk::CommandBuffer;
+
+    const TYPE: vk::DebugReportObjectTypeEXT = vk::DEBUG_REPORT_OBJECT_TYPE_COMMAND_BUFFER_EXT;
 
     #[inline]
     fn internal_object(&self) -> vk::CommandBuffer {