+++ vulkano/src/framebuffer/sys.rs
@@ -496,6 +496,8 @@ pub struct RenderPassSys<'a>(vk::RenderPass, PhantomData<&'a ()>);
 unsafe impl<'a> VulkanObject for RenderPassSys<'a> {
     type Object = vk::RenderPass;
 
+    const TYPE: vk::DebugReportObjectTypeEXT = vk::DEBUG_REPORT_OBJECT_TYPE_RENDER_PASS_EXT;
+
     #[inline]
     fn internal_object(&self) -> vk::RenderPass {
         self.0