+++ vulkano/src/pipeline/graphics_pipeline/mod.rs
@@ -302,6 +302,8 @@ unsafe impl<C, Mv, L, Rp> RenderPassDescClearValues<C> for GraphicsPipeline<Mv, 
 unsafe impl<Mv, L, Rp> VulkanObject for GraphicsPipeline<Mv, L, Rp> {
     type Object = vk::Pipeline;
 
+    const TYPE: vk::DebugReportObjectTypeEXT = vk::DEBUG_REPORT_OBJECT_TYPE_PIPELINE_EXT;
+
     #[inline]
     fn internal_object(&self) -> vk::Pipeline {
         self.inner.pipeline
@@ -476,6 +478,8 @@ pub struct GraphicsPipelineSys<'a>(vk::Pipeline, PhantomData<&'a ()>);
 
 unsafe impl<'a> VulkanObject for GraphicsPipelineSys<'a> {
     type Object = vk::Pipeline;
+
+    const TYPE: vk::DebugReportObjectTypeEXT = vk::DEBUG_REPORT_OBJECT_TYPE_PIPELINE_EXT;
 
     #[inline]
     fn internal_object(&self) -> vk::Pipeline {