+++ vulkano/src/query/mod.rs
@@ -114,6 +114,8 @@ impl UnsafeQueryPool {
 unsafe impl VulkanObject for UnsafeQueryPool {
     type Object = vk::QueryPool;
 
+    const TYPE: vk::DebugReportObjectTypeEXT = vk::DEBUG_REPORT_OBJECT_TYPE_QUERY_POOL_EXT;
+
     #[inline]
     fn internal_object(&self) -> vk::QueryPool {
         self.pool